Move shorten_fname{,1,s} and shorten_filenames from fileio.c
